chore: move docker namespace calculation to docker file
